setup models for db

+++ Models/ArtistLike.cs
@@ -2,10 +2,10 @@ namespace MusicDb.Models {
 
   public class ArtistLike : BaseEntity {
 
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; }
 
-    public int ArtistId { get; set; }
+    public long ArtistId { get; set; }
     public Artist Artist { get; set; }
 
   }
